requirements.txt
absl-py==2.3.1
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
apricot-select==0.6.1
async-timeout==5.0.1
attrs==25.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.14.3
blis==0.4.1
blobfile==3.1.0
catalogue==1.0.2
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.7
click==8.1.8
contextlib2==21.6.0
contourpy==1.1.1
cryptography==46.0.6
cycler==0.12.1
cymem==2.0.11
Cython==0.29.32
datasets==3.1.0
decorator==4.4.2
diffusers==0.36.0
dill==0.3.8
easydict==1.13
einops==0.8.1
ema-pytorch==0.7.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz#sha256=ea8c87848b4a97ced174919e08c00c7888a30495ace38d281d455ee270da2c12
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fairscale==0.4.13
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-auth-oauthlib==1.0.0
graphviz==0.20.3
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
image-reward==1.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912944151/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240558244/work
ipython_genutils @ file:///home/conda/feedstock_root/build_artifacts/ipython_genutils_1716278396992/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.6
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257397447/work
kiwisolver==1.4.7
labelImg==1.8.6
langcodes==3.4.1
language_data==1.4.0
llvmlite==0.41.1
lxml==6.1.0
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
ml_collections==0.1.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.15
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==2.8.7
nose==1.3.7
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.3.1
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.13.0.92
packaging==26.0
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathlib_abc==0.1.1
pathy==0.11.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
plac==1.1.3
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pomegranate==1.1.2
preshed==3.0.12
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1717574682739/work
propcache==0.2.0
protobuf==5.29.6
psutil==7.2.2
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==17.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==1.8.2
pydantic_core==2.27.2
pyemd==1.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.15.0
pyrallis==0.3.1
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1666828534876/work
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.57.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.4.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.7
spacy==2.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thinc==7.4.1
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.13.3
torch==2.3.0+cu118
torch_fidelity==0.4.0
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1666788592778/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.27.4
triton==2.3.0
typer==0.4.2
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.2.3
wandb==0.24.2
wasabi==0.10.1
wcwidth==0.2.14
Werkzeug==3.0.6
xxhash==3.6.0
yarl==1.15.2
zipp==3.20.2
